Moved WriteoutOptions to Potentials
