+++ modules/algorithms/src/main/groovy/ffx/algorithms/groovy/DynamicsOpenMM.groovy
@@ -42,15 +42,13 @@ import org.apache.commons.io.FilenameUtils
 import ffx.algorithms.cli.AlgorithmsScript
 import ffx.algorithms.cli.BarostatOptions
 import ffx.algorithms.cli.DynamicsOptions
-import ffx.algorithms.cli.WriteoutOptions
+import ffx.potential.cli.WriteoutOptions
 import ffx.algorithms.dynamics.MolecularDynamics
 import ffx.algorithms.dynamics.MolecularDynamicsOpenMM
 import ffx.numerics.Potential
 import ffx.potential.ForceFieldEnergy
 import ffx.potential.ForceFieldEnergyOpenMM
 import ffx.potential.MolecularAssembly
-import ffx.potential.parameters.ForceField
-
 import picocli.CommandLine.Command
 import picocli.CommandLine.Mixin
 import picocli.CommandLine.Option
